Use editor's shared audio context in sound library instead of scratch-audio shared context
Vanilla scratch-audio uses a shared context. We don't want that
for this because we want to be able to suspend() the vm's audio
context without it resuming when someone opens the sound
library later.

We removed that caching, but now we would be making a new audio
context every time. Instead we can just use the shared one that we
already have.

Related to https://github.com/ScratchAddons/ScratchAddons/issues/6847